reworked the way help is presented and handled in the code, added feature to also directly request a command information, restructured administration, utility and channel management into partial classes, for easier management, fixes #248
