+++ src/OnePlusBot/Modules/Administration/Administration.cs
@@ -1,0 +1,13 @@
+using Discord.Commands;
+
+
+namespace OnePlusBot.Modules.Administration
+{
+    [
+      Summary("Utilities for administrating the server: User moderation/metrics and bot configuration")
+    ]
+    public partial class Administration : ModuleBase<SocketCommandContext>
+    {
+        
+    }
+}
