+++ src/OnePlusBot/Modules/Channels/Channels.cs
@@ -1,0 +1,14 @@
+using Discord.Commands;
+using Discord.Addons.Interactive;
+
+
+namespace OnePlusBot.Modules.Channels
+{
+    [
+      Summary("Module containing commands to configure channel groups and post targets")
+    ]
+    public partial class Channels : InteractiveBase<SocketCommandContext>
+    {
+        
+    }
+}
