+++ src/OnePlusBot/Modules/Utility/Utility.cs
@@ -1,0 +1,12 @@
+using Discord.Commands;
+
+namespace OnePlusBot.Modules.Utility
+{
+
+    [
+      Summary("Module containing utility commands. Examples are: showavatar, userinfo, suggest")
+    ]
+    public partial class Utility : ModuleBase<SocketCommandContext>
+    {
+    }
+}
